Migrated sequences
Migrated the sequences to the new extension system. Made a couple large charges to hopefully increase response.
Each sequence use to spawn a thread and block for their `delay` and `duration`. Instead there is no blocking and now just checking for time elapsed. This allows less threads for sequence and enabled them to be a `component` instead of `worker`.