Updated Event format to be more consistent
Updated the event format used across the event bus. Now events will expect all extra data to be put inside a `data` attribute on the dict passed in. This makes things more uniform and allows an expected format when parsing event data. If there is extra data passed in not inside the `data` key then MudPi will package the data for you and place it inisde this key for legacy support.